Rename `pyproject` -> `pyproject_toml` for clarity
