+++ src/usethis/_core/readme.py
@@ -1,8 +1,8 @@
 from pathlib import Path
 
 from usethis._console import box_print, tick_print
-from usethis._integrations.pyproject.errors import PyProjectTOMLError
-from usethis._integrations.pyproject.name import get_description, get_name
+from usethis._integrations.pyproject_toml.errors import PyProjectTOMLError
+from usethis._integrations.pyproject_toml.name import get_description, get_name
 from usethis._integrations.uv.init import ensure_pyproject_toml
