+++ src/usethis/_core/show.py
@@ -2,7 +2,7 @@ import typer
 
 from usethis._config import usethis_config
 from usethis._console import err_print
-from usethis._integrations.pyproject.name import get_name
+from usethis._integrations.pyproject_toml.name import get_name
 from usethis._integrations.sonarqube.config import get_sonar_project_properties
 from usethis._integrations.uv.init import ensure_pyproject_toml
 from usethis.errors import UsethisError
