+++ src/usethis/_integrations/pyproject_toml/requires_python.py
@@ -1,7 +1,7 @@
 from packaging.specifiers import SpecifierSet
 from pydantic import TypeAdapter
 
-from usethis._integrations.pyproject.io_ import read_pyproject_toml
+from usethis._integrations.pyproject_toml.io_ import read_pyproject_toml
 
 
 class MissingRequiresPythonError(Exception):
