+++ src/usethis/_integrations/sonarqube/config.py
@@ -4,7 +4,7 @@ from pathlib import Path
 from pydantic import TypeAdapter
 
 from usethis._integrations.project.layout import get_source_dir_str
-from usethis._integrations.pyproject.core import get_config_value
+from usethis._integrations.pyproject_toml.core import get_config_value
 from usethis._integrations.python.version import get_python_version
 from usethis._integrations.sonarqube.errors import (
     CoverageReportConfigNotFoundError,
