+++ src/usethis/_interface/badge.py
@@ -9,7 +9,7 @@ from usethis._core.badge import (
     remove_pypi_badge,
     remove_ruff_badge,
 )
-from usethis._integrations.pyproject.io_ import pyproject_toml_io_manager
+from usethis._integrations.pyproject_toml.io_ import pyproject_toml_io_manager
 
 app = typer.Typer(help="Add badges to the top of the README.md file.")
