+++ src/usethis/_interface/readme.py
@@ -3,7 +3,7 @@ import typer
 from usethis._config import quiet_opt, usethis_config
 from usethis._core.badge import add_pre_commit_badge, add_ruff_badge
 from usethis._core.readme import add_readme
-from usethis._integrations.pyproject.io_ import pyproject_toml_io_manager
+from usethis._integrations.pyproject_toml.io_ import pyproject_toml_io_manager
 from usethis._tool import PreCommitTool, RuffTool
