+++ tests/usethis/_integrations/pyproject_toml/test_project.py
@@ -2,12 +2,12 @@ from pathlib import Path
 
 import pytest
 
-from usethis._integrations.pyproject.errors import (
+from usethis._integrations.pyproject_toml.errors import (
     PyProjectTOMLNotFoundError,
     PyProjectTOMLProjectSectionError,
 )
-from usethis._integrations.pyproject.io_ import pyproject_toml_io_manager
-from usethis._integrations.pyproject.project import get_project_dict
+from usethis._integrations.pyproject_toml.io_ import pyproject_toml_io_manager
+from usethis._integrations.pyproject_toml.project import get_project_dict
 from usethis._test import change_cwd
