+++ tests/usethis/_integrations/ruff/test_rules.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 import pytest
 
-from usethis._integrations.pyproject.io_ import (
+from usethis._integrations.pyproject_toml.io_ import (
     PyProjectTOMLNotFoundError,
     pyproject_toml_io_manager,
 )
